Support Eprints
Article on arxiv are advocated to be cited as

Moseley, C. (2014). “Simpler proof of the theorem by Pusey, Barrett,
and Rudolph on the reality of the quantum state”. In: ArXiv e-prints.
arXiv: 1401.0026 [quant-ph].

To this effect we introduce three new fields on the bibliography.
In this example:
        eprint_archive: arXiv
        eprint_number: 1401.0026
        eprint_class: quant-ph

This maps the Biblatex fields Archiveprefix, Eprint, and Primaryclass,
respectively.