Simplify the Bell theorem without math.
I use cards instead of tokens and insist on the fact that we gain
knowledge only through the detectors. My tokens with two faces X and Y
gave the impression that the problem could be with the physical reality
of the system only, making it impossible to discuss the detector.